Fix DooD for Windows
When mounting the socket Docker wants some sort of weird Windows-like
path, to which I couldn't find an acceptable version which also showed
the contents of this folder. Both the HyperV and the WSL2 backend fail
at this in a different way

Luckily Docker has a workaround: in the settings you can enable exposing
the daemon on a TCP socket. This is a small security hole, but it does
work: you can just feed it a windows path and it will start the
container with the correct mount.

I'm open for better solutions