Remove .NET Core 3.1 image, cache NuGet deps
This commit combines 3 strategies to save space while building:

Exiled.Patcher used .NET Core 3.1, however we override this to .NET 5.0:
we already have that image for APublicizer and this allows us to remove
3.1 altogether, saving out 700 MB

.local/share/NuGet is a cache directory, we now mount this seperately, which
speeds up rebuilds of EXPEL and saves 100MB as this cache used to be
stored by both the patcher and the main exiled build

The main EXILED build used to restore all projects, even the .NET Core
projects (Installer and Patcher), while the latter are of course not
built. We now only restore the .NET Framework projects, which saves
around 350 MB. Yes, the dependencies for the Installer and Patcher are
quite large.